+++ x/cronos/keeper/msg_server.go
@@ -4,7 +4,9 @@ import (
 	"context"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/crypto-org-chain/cronos/x/cronos/types"
+	"github.com/ethereum/go-ethereum/common"
 )
 
 type msgServer struct {
@@ -54,4 +56,17 @@ func (k msgServer) TransferTokens(goCtx context.Context, msg *types.MsgTransferT
 		)},
 	)
 	return &types.MsgTransferTokensResponse{}, nil
+}
+
+// UpdateTokenMapping implements the grpc method
+func (k msgServer) UpdateTokenMapping(goCtx context.Context, msg *types.MsgUpdateTokenMapping) (*types.MsgUpdateTokenMappingResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	admin := k.Keeper.GetParams(ctx).CronosAdmin
+	// if admin is empty, no sender could be equal to it
+	if admin != msg.Sender {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "msg sender is authorized")
+	}
+	// msg is already validated
+	k.Keeper.SetExternalContractForDenom(ctx, msg.Denom, common.HexToAddress(msg.Contract))
+	return &types.MsgUpdateTokenMappingResponse{}, nil
 }